refactor: move folders layout
