+++ frontend/src/components/Recipes/RecipesLayout.tsx
@@ -1,0 +1,9 @@
+import Recipes from './Content'
+
+export default function RecipesLayout() {
+    return (
+        <div>
+            <Recipes />
+        </div>
+    )
+}
